tools/dump_tool: Refactoring of dump_tool code
- Add header file dump_tool.h for common definitions and macros.
- Add dump_tool_utils.c for utility functions used by dump tool.
